requirements.txt
black==19.10b0
eth-brownie>=1.13.2,<2.0.0
flake8==3.8.4
isort==5.7.0
flaky>=3.7.0
